Error Prone: Fix EqualsGetClass violations in GameChooserEntry
GameChooserEntry has deep integration with other subsystems, which
explains why there are unit tests that mock it to avoid the integration.
That's kind of a smell, so I transformed GameChooserEntry into an
interface and extracted the existing implementation into the new
DefaultGameChooserEntry class.

Once that was complete, it was a simple matter to fix the equals()
implementation to conform to the Error Prone rule.
